requirements.txt
click==8.1.8
pillow==11.2.1
